refactor(p2p): forging index should be calculated correctly for dynamic block times
